requirements.txt
FastAPI
uvicorn
asyncpg
SQLAlchemy
httpx
python-dotenv
aiosqlite
greenlet
mcp
pydantic
pydantic-settings 